Naive Implementation of Provenance in Souffle (#418)
Implements a naive provenance system for Souffle. Records reference numbers are used as proof labels to generate proof trees. The provenance system has been implemented as an Ast-Transformation converting tuples to records.  A simple provenance text-user interface is provided for the compiler and interpreter. The option -t enables the Ast-Transformation and the provenance text user interface at the end of execution. The text user interface is controlled via simple commands like "explain" to explain the proof tree of a tuple and exit for terminating the text user interface.